checking role for user

+++ app/Http/Controllers/ProductController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreProductRequest;
 use App\Models\Product;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Models\Role;
 
 class ProductController extends Controller
 {
@@ -19,31 +22,48 @@ class ProductController extends Controller
     }
     public function store(StoreProductRequest $request)
     {
-        $product = new Product();
-        $product = $product->create($request->toArray());
-        return response()->json(['message' => 'محصول با موفقیت ایجاد شد', 'product' => $product]);
+        $user = new User();
+        $user = $user->find(Auth::id());
+        if($user->hasRole('admin')) {
+            $product = new Product();
+            $product = $product->create($request->toArray());
+            return response()->json(['message' => 'محصول با موفقیت ایجاد شد', 'product' => $product]);
+        }
     }
     public function update($request, $id)
     {
-        $product = new Product();
-        $product = $product->find($id);
-        $product->update($request->validated());
-        return response()->json(['message' => 'محصول با موفقیت به روز رسانی شد', 'product' => $product]);
+        $user = new User();
+        $user = find(Auth::id());
+        if ($user->hasRole('admin'))
+        {
+            $product = new Product();
+            $product = $product->find($id);
+            $product->update($request->validated());
+            return response()->json(['message' => 'محصول با موفقیت به روز رسانی شد', 'product' => $product]);
+        }
     }
 
     public function destroy($id)
     {
-        $product = new Product();
-        $product = $product->find($id);
-        $product->delete();
-        return response()->json(['محصول با موفقیت حذف شد ']);
+        $user = new User();
+        $user = $user->find(Auth::id());
+        if ($user->hasRole('admin')) {
+            $product = new Product();
+            $product = $product->find($id);
+            $product->delete();
+            return response()->json(['محصول با موفقیت حذف شد ']);
+        }
     }
 
     public function restore($id)
     {
-        $product = new Product();
-        $product = $product->withTrashed()->find($id);
-        $product->restore();
-        return response()->json(['message' => ' محصول با موفقیت بازگردانده شد ', 'product' => $product]);
+        $user = new User();
+        $user = find(Auth::id());
+        if ($user->hasRole('admin')) {
+            $product = new Product();
+            $product = $product->withTrashed()->find($id);
+            $product->restore();
+            return response()->json(['message' => ' محصول با موفقیت بازگردانده شد ', 'product' => $product]);
+        }
     }
 }
